+++ apps/mmessier/app/experience/[experienceID]/page.tsx
@@ -11,16 +11,24 @@ type ExperienceDetailsProps = {
 const ExperienceDetails = async ({
   params: { experienceID },
 }: ExperienceDetailsProps) => {
-  const experience = await getExperience(experienceID);
-  return <ExperienceDetailsClient experience={experience} />;
+  const experienceResponse = await getExperience(experienceID);
+  if (experienceResponse?.experience) {
+    return (
+      <ExperienceDetailsClient experience={experienceResponse?.experience} />
+    );
+  }
+  return null;
 };
 
 export default ExperienceDetails;
 
 // This changes the [experienceID] route from a λ (SSR) to an ● (SSG) components, which means they are generated in-advance by the server
 export const generateStaticParams = async () => {
-  const experiences = await getExperiences();
-  return experiences.map((experience) => ({
-    experienceID: experience.id,
-  }));
+  const experiencesResponse = await getExperiences();
+  if (experiencesResponse?.experiences) {
+    return experiencesResponse.experiences.map((experience) => ({
+      experienceID: experience.id,
+    }));
+  }
+  return null;
 };
